Fixing documentation issue
