+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/site/public/locales/en.json
@@ -91,9 +91,7 @@
   "Apis.Details.Overview.sdk.generation.title": "SDK Generation",
   "Apis.Details.Overview.subscribe.to.application": "Generate Credentials",
   "Apis.Details.Overview.subscribe.to.application.btn": "Subscribe to an Application",
-  "Apis.Details.Overview.subscribe.to.application.content.Application": "Application",
-  "Apis.Details.Overview.subscribe.to.application.content.Applications": "Applications",
-  "Apis.Details.Overview.subscriptions": "Subscriptions",
+  "Apis.Details.Overview.subscriptions": "Subscriptions >>",
   "Apis.Details.Overview.view.credentials": "View Credentials",
   "Apis.Details.Sdk.no.lanuages": "No languages are configured.",
   "Apis.Details.Subscribe.applications.title": "Applications",
